requirements.txt
absl-py==2.2.2
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.9
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
keras==2.14.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.8
numpy==1.24.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
scipy==1.15.2
six==1.17.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.14.0
tensorflow-gan==2.1.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.14.0
tensorflow-metal==1.2.0
tensorflow-probability==0.22.0
termcolor==3.0.1
tqdm==4.67.1
typing_extensions==4.5.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.14.1
